[fixed] City radius
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2545 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/city/CityManager.h
@@ -160,7 +160,7 @@ class CityManagerImplementation : public ManagedServiceImplementation, public Lo
 	ManagedWeakReference<Zone* > zone;
 
 public:
-	static Vector<byte> radiusPerRank;
+	static Vector<unsigned short> radiusPerRank;
 
 	static Vector<byte> citizensPerRank;
